backend/requirements.txt
aniso8601
bcrypt
blinker
cffi
click
Flask
Flask-Bcrypt
Flask-Mail
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
greenlet
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
marshmallow
packaging
psycopg-binary
psycopg2
pycparser
PyJWT
pyparsing
python-dotenv
pytz
six
SQLAlchemy
Werkzeug
zipp
pyGithub
linkedin-api~=2.0.0a
pandas
requests
beautifulsoup4